Server/Model-API/requirements.txt
Flask==2.0.2
Flask-Cors
moviepy==1.0.3
flask-socketio==5.1.1
Werkzeug==2.0.2
pytube==11.0.1
opencv-python==4.5.4.60 
torch==1.10.1 
numpy==1.21.5 
Pillow==9.0.1
soundfile==0.11.1
librosa==1.1.1
ortools==9.0.9048
